input: fingerprint: goodix_ta: rewrite for performance goodix's implementation is quite generic at it's core and so, there was room for improvement without hurting the basic functions debugging and unsupported features are removed and also removes the need for a display notifier and a wakelock as our SoC can receive interrupts from sleep and respond just fine.
generally, mutexes and other bogus checks have also been removed
and indents have been fixed, driver should be a lot more readable
now that everything's in one place.

this rewrite can be adapted to work on the majority of goodix sensors
(with minor edits, say, pinctrl or gpio names or new regulators and such.)

tested to be working on 3208, 3258, and fod variants
fp sensor performance had improved by over 50% on 8953 with the same rewrite.

DuckNote: for beryllium
* move common function calls under gf_setup or cleanup
* move defines just before they're used
* rework netlink sender logic
* rework irq handler
* fixup mutexes and function call ordering ( kernel panic fix )

There is about 40% Performance improvement on beryllium
also rename it to stripped to keep original goodix driver.

Signed-off-by: duckyduckg <duckyduckg65@gmail.com>